requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
Pillow==10.2.0
psycopg2-binary==2.9.9   # If using PostgreSQL (remove if not needed)
mysqlclient==2.2.4       # If using MySQL (remove if not needed)
python-decouple==3.8     # For managing environment variables
